requirements.txt
tensorflow 
Pillow
numpy 
fastapi
uvicorn
python-multipart
